bump api version to v2
