+++ api-server/stack-test-suite/tests/v2/address.rs
@@ -13,10 +13,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{collections::BTreeMap, sync::RwLock};
+use std::sync::RwLock;
 
-use api_web_server::CachedValues;
-use common::{chain::UtxoOutPoint, primitives::time::get_time};
+use api_web_server::{api::json_helpers::amount_to_json, CachedValues};
+use common::primitives::time::get_time;
 
 use crate::DummyRPC;
 
@@ -24,7 +24,7 @@ use super::*;
 
 #[tokio::test]
 async fn invalid_address() {
-    let (task, response) = spawn_webserver("/api/v1/address/invalid-address/all-utxos").await;
+    let (task, response) = spawn_webserver("/api/v2/address/invalid-address").await;
 
     assert_eq!(response.status(), 400);
 
@@ -48,16 +48,14 @@ async fn address_not_found(#[case] seed: Seed) {
     let destination = Destination::PublicKeyHash(PublicKeyHash::from(&public_key));
     let address = Address::<Destination>::new(&chain_config, &destination).unwrap();
 
-    let (task, response) =
-        spawn_webserver(&format!("/api/v1/address/{}/all-utxos", address.get())).await;
+    let (task, response) = spawn_webserver(&format!("/api/v2/address/{}", address.get())).await;
 
-    assert_eq!(response.status(), 200);
+    assert_eq!(response.status(), 404);
 
     let body = response.text().await.unwrap();
     let body: serde_json::Value = serde_json::from_str(&body).unwrap();
-    let utxos = body.as_array().unwrap();
 
-    assert!(utxos.is_empty());
+    assert_eq!(body["error"].as_str().unwrap(), "Address not found");
 
     task.abort();
 }
@@ -66,11 +64,11 @@ async fn address_not_found(#[case] seed: Seed) {
 #[trace]
 #[case(Seed::from_entropy())]
 #[tokio::test]
-async fn multiple_utxos_to_single_address(#[case] seed: Seed) {
+async fn multiple_outputs_to_single_address(#[case] seed: Seed) {
     let listener = TcpListener::bind("127.0.0.1:0").unwrap();
     let addr = listener.local_addr().unwrap();
 
-    let (tx, rx) = tokio::sync::oneshot::channel::<[(String, serde_json::Value); 2]>();
+    let (tx, rx) = tokio::sync::oneshot::channel();
 
     let task = tokio::spawn(async move {
         let web_server_state = {
@@ -91,7 +89,7 @@ async fn multiple_utxos_to_single_address(#[case] seed: Seed) {
                 let alice_address =
                     Address::<Destination>::new(&chain_config, &alice_destination).unwrap();
                 let mut alice_balance = Amount::from_atoms(1_000_000);
-                let mut alice_utxos = BTreeMap::new();
+                let mut alice_transaction_history: Vec<Id<Transaction>> = vec![];
 
                 let (_bob_sk, bob_pk) =
                     PrivateKey::new_from_rng(&mut rng, KeyKind::Secp256k1Schnorr);
@@ -99,7 +97,9 @@ async fn multiple_utxos_to_single_address(#[case] seed: Seed) {
                 let bob_destination = Destination::PublicKeyHash(PublicKeyHash::from(&bob_pk));
                 let bob_address =
                     Address::<Destination>::new(&chain_config, &bob_destination).unwrap();
-                let mut bob_utxos = BTreeMap::new();
+                let mut bob_balance = Amount::ZERO;
+                let mut bob_locked_balance = Amount::ZERO;
+                let mut bob_transaction_history: Vec<Id<Transaction>> = vec![];
 
                 // setup initial transaction
 
@@ -117,10 +117,8 @@ async fn multiple_utxos_to_single_address(#[case] seed: Seed) {
                     .add_output(previous_tx_out.clone())
                     .build();
 
+                alice_transaction_history.push(transaction.transaction().get_id());
                 let previous_transaction_id = transaction.transaction().get_id();
-                let utxo =
-                    UtxoOutPoint::new(OutPointSourceId::Transaction(previous_transaction_id), 0);
-                alice_utxos.insert(utxo, previous_tx_out.clone());
 
                 let mut previous_witness = InputWitness::Standard(
                     StandardInputSignature::produce_uniparty_signature_for_input(
@@ -146,9 +144,16 @@ async fn multiple_utxos_to_single_address(#[case] seed: Seed) {
 
                 let random_coin_amount1 = rng.gen_range(1..10);
                 let random_coin_amount2 = rng.gen_range(1..10);
+                let random_coin_amount3 = rng.gen_range(1..10);
 
                 alice_balance = (alice_balance - Amount::from_atoms(random_coin_amount1)).unwrap();
                 alice_balance = (alice_balance - Amount::from_atoms(random_coin_amount2)).unwrap();
+                alice_balance = (alice_balance - Amount::from_atoms(random_coin_amount3)).unwrap();
+
+                bob_balance = (bob_balance + Amount::from_atoms(random_coin_amount1)).unwrap();
+                bob_balance = (bob_balance + Amount::from_atoms(random_coin_amount2)).unwrap();
+                bob_locked_balance =
+                    (bob_locked_balance + Amount::from_atoms(random_coin_amount3)).unwrap();
 
                 let alice_tx_out =
                     TxOutput::Transfer(OutputValue::Coin(alice_balance), alice_destination.clone());
@@ -158,8 +163,13 @@ async fn multiple_utxos_to_single_address(#[case] seed: Seed) {
                     bob_destination.clone(),
                 );
 
-                let bob_tx_out2 = TxOutput::LockThenTransfer(
+                let bob_tx_out2 = TxOutput::Transfer(
                     OutputValue::Coin(Amount::from_atoms(random_coin_amount2)),
+                    bob_destination.clone(),
+                );
+
+                let bob_tx_out3 = TxOutput::LockThenTransfer(
+                    OutputValue::Coin(Amount::from_atoms(random_coin_amount3)),
                     bob_destination.clone(),
                     OutputTimeLock::ForBlockCount(10),
                 );
@@ -175,27 +185,11 @@ async fn multiple_utxos_to_single_address(#[case] seed: Seed) {
                     .add_output(alice_tx_out.clone())
                     .add_output(bob_tx_out1.clone())
                     .add_output(bob_tx_out2.clone())
+                    .add_output(bob_tx_out3.clone())
                     .build();
 
-                let utxo = UtxoOutPoint::new(
-                    OutPointSourceId::Transaction(transaction.transaction().get_id()),
-                    0,
-                );
-                alice_utxos.insert(utxo, alice_tx_out);
-                bob_utxos.insert(
-                    UtxoOutPoint::new(
-                        OutPointSourceId::Transaction(transaction.transaction().get_id()),
-                        1,
-                    ),
-                    bob_tx_out1,
-                );
-                bob_utxos.insert(
-                    UtxoOutPoint::new(
-                        OutPointSourceId::Transaction(transaction.transaction().get_id()),
-                        2,
-                    ),
-                    bob_tx_out2,
-                );
+                alice_transaction_history.push(transaction.transaction().get_id());
+                bob_transaction_history.push(transaction.transaction().get_id());
 
                 previous_witness = InputWitness::Standard(
                     StandardInputSignature::produce_uniparty_signature_for_input(
@@ -227,27 +221,19 @@ async fn multiple_utxos_to_single_address(#[case] seed: Seed) {
                 _ = tx.send([
                     (
                         alice_address.get().to_string(),
-                        alice_utxos
-                            .into_iter()
-                            .map(|utxo| {
-                                json!({
-                                "outpoint": utxo.0,
-                                "utxo": txoutput_to_json(&utxo.1, &chain_config, &TokenDecimals::Single(None))})
-                            })
-                            .collect::<Vec<_>>()
-                            .into(),
+                        json!({
+                        "coin_balance": amount_to_json(alice_balance, chain_config.coin_decimals()),
+                        "locked_coin_balance": amount_to_json(Amount::ZERO, chain_config.coin_decimals()),
+                        "transaction_history": alice_transaction_history,
+                                }),
                     ),
                     (
                         bob_address.to_string(),
-                        bob_utxos
-                            .into_iter()
-                            .map(|utxo| {
-                                json!({
-                                "outpoint": utxo.0,
-                                "utxo": txoutput_to_json(&utxo.1, &chain_config, &TokenDecimals::Single(None))})
-                            })
-                            .collect::<Vec<_>>()
-                            .into(),
+                        json!({
+                        "coin_balance": amount_to_json(bob_balance, chain_config.coin_decimals()),
+                        "locked_coin_balance": amount_to_json(bob_locked_balance, chain_config.coin_decimals()),
+                        "transaction_history": bob_transaction_history,
+                                }),
                     ),
                 ]);
 
@@ -287,8 +273,8 @@ async fn multiple_utxos_to_single_address(#[case] seed: Seed) {
         web_server(listener, web_server_state, true).await
     });
 
-    for (address, expected) in rx.await.unwrap() {
-        let url = format!("/api/v1/address/{address}/all-utxos");
+    for (address, expected_balance) in rx.await.unwrap() {
+        let url = format!("/api/v2/address/{address}");
 
         // Given that the listener port is open, this will block until a
         // response is made (by the web server, which takes the listener
@@ -306,8 +292,7 @@ async fn multiple_utxos_to_single_address(#[case] seed: Seed) {
         let body = response.text().await.unwrap();
         let body: serde_json::Value = serde_json::from_str(&body).unwrap();
 
-        eprintln!("asd\n\n");
-        assert_eq!(body, expected);
+        assert_eq!(body, expected_balance);
     }
 
     task.abort();
@@ -317,11 +302,11 @@ async fn multiple_utxos_to_single_address(#[case] seed: Seed) {
 #[trace]
 #[case(Seed::from_entropy())]
 #[tokio::test]
-async fn ok(#[case] seed: Seed) {
+async fn test_unlocking_for_locked_utxos(#[case] seed: Seed) {
     let listener = TcpListener::bind("127.0.0.1:0").unwrap();
     let addr = listener.local_addr().unwrap();
 
-    let (tx, rx) = tokio::sync::oneshot::channel::<[(String, serde_json::Value); 2]>();
+    let (tx, rx) = tokio::sync::oneshot::channel();
 
     let task = tokio::spawn(async move {
         let web_server_state = {
@@ -342,7 +327,7 @@ async fn ok(#[case] seed: Seed) {
                 let alice_address =
                     Address::<Destination>::new(&chain_config, &alice_destination).unwrap();
                 let mut alice_balance = Amount::from_atoms(1_000_000);
-                let mut alice_utxos = BTreeMap::new();
+                let mut alice_transaction_history: Vec<Id<Transaction>> = vec![];
 
                 let (_bob_sk, bob_pk) =
                     PrivateKey::new_from_rng(&mut rng, KeyKind::Secp256k1Schnorr);
@@ -350,11 +335,13 @@ async fn ok(#[case] seed: Seed) {
                 let bob_destination = Destination::PublicKeyHash(PublicKeyHash::from(&bob_pk));
                 let bob_address =
                     Address::<Destination>::new(&chain_config, &bob_destination).unwrap();
-                let mut bob_utxos = BTreeMap::new();
+                let mut bob_balance = Amount::ZERO;
+                let mut bob_locked_balance = Amount::ZERO;
+                let mut bob_transaction_history: Vec<Id<Transaction>> = vec![];
 
                 // setup initial transaction
 
-                let mut previous_tx_out =
+                let previous_tx_out =
                     TxOutput::Transfer(OutputValue::Coin(alice_balance), alice_destination.clone());
 
                 let transaction = TransactionBuilder::new()
@@ -368,14 +355,8 @@ async fn ok(#[case] seed: Seed) {
                     .add_output(previous_tx_out.clone())
                     .build();
 
-                let mut previous_transaction_id = transaction.transaction().get_id();
-                alice_utxos.insert(
-                    UtxoOutPoint::new(
-                        OutPointSourceId::Transaction(transaction.transaction().get_id()),
-                        0,
-                    ),
-                    previous_tx_out.clone(),
-                );
+                alice_transaction_history.push(transaction.transaction().get_id());
+                let previous_transaction_id = transaction.transaction().get_id();
 
                 let mut previous_witness = InputWitness::Standard(
                     StandardInputSignature::produce_uniparty_signature_for_input(
@@ -397,11 +378,255 @@ async fn ok(#[case] seed: Seed) {
                     .unwrap()
                     .block_id()];
 
-                for _ in 0..rng.gen_range(1..2) {
+                // Generate two outputs for a single transaction
+
+                let random_coin_amount1 = rng.gen_range(1..10);
+                let random_coin_amount2 = rng.gen_range(1..10);
+                let random_coin_amount3 = rng.gen_range(1..10);
+
+                alice_balance = (alice_balance - Amount::from_atoms(random_coin_amount1)).unwrap();
+                alice_balance = (alice_balance - Amount::from_atoms(random_coin_amount2)).unwrap();
+                alice_balance = (alice_balance - Amount::from_atoms(random_coin_amount3)).unwrap();
+
+                bob_balance = (bob_balance + Amount::from_atoms(random_coin_amount1)).unwrap();
+                bob_balance = (bob_balance + Amount::from_atoms(random_coin_amount2)).unwrap();
+                bob_locked_balance =
+                    (bob_locked_balance + Amount::from_atoms(random_coin_amount3)).unwrap();
+
+                let alice_tx_out =
+                    TxOutput::Transfer(OutputValue::Coin(alice_balance), alice_destination.clone());
+
+                let bob_tx_out1 = TxOutput::Transfer(
+                    OutputValue::Coin(Amount::from_atoms(random_coin_amount1)),
+                    bob_destination.clone(),
+                );
+
+                let bob_tx_out2 = TxOutput::Transfer(
+                    OutputValue::Coin(Amount::from_atoms(random_coin_amount2)),
+                    bob_destination.clone(),
+                );
+
+                let bob_tx_out3 = TxOutput::LockThenTransfer(
+                    OutputValue::Coin(Amount::from_atoms(random_coin_amount3)),
+                    bob_destination.clone(),
+                    OutputTimeLock::ForBlockCount(1),
+                );
+
+                let transaction = TransactionBuilder::new()
+                    .add_input(
+                        TxInput::from_utxo(
+                            OutPointSourceId::Transaction(previous_transaction_id),
+                            0,
+                        ),
+                        previous_witness.clone(),
+                    )
+                    .add_output(alice_tx_out.clone())
+                    .add_output(bob_tx_out1.clone())
+                    .add_output(bob_tx_out2.clone())
+                    .add_output(bob_tx_out3.clone())
+                    .build();
+
+                alice_transaction_history.push(transaction.transaction().get_id());
+                bob_transaction_history.push(transaction.transaction().get_id());
+
+                previous_witness = InputWitness::Standard(
+                    StandardInputSignature::produce_uniparty_signature_for_input(
+                        &alice_sk,
+                        SigHashType::try_from(SigHashType::ALL).unwrap(),
+                        alice_destination.clone(),
+                        &transaction,
+                        &[Some(&previous_tx_out)],
+                        0,
+                    )
+                    .unwrap(),
+                );
+
+                let signed_transaction = SignedTransaction::new(
+                    transaction.transaction().clone(),
+                    vec![previous_witness.clone()],
+                )
+                .unwrap();
+
+                chainstate_block_ids.push(
+                    *tf.make_block_builder()
+                        .add_transaction(signed_transaction)
+                        .build_and_process()
+                        .unwrap()
+                        .unwrap()
+                        .block_id(),
+                );
+
+                // Add an empty block that to wait for the time lock to pass
+                chainstate_block_ids.push(
+                    *tf.make_block_builder().build_and_process().unwrap().unwrap().block_id(),
+                );
+
+                _ = tx.send([
+                    (
+                        alice_address.get().to_string(),
+                        json!({
+                        "coin_balance": amount_to_json(alice_balance, chain_config.coin_decimals()),
+                        "locked_coin_balance": amount_to_json(Amount::ZERO, chain_config.coin_decimals()),
+                        "transaction_history": alice_transaction_history,
+                                }),
+                    ),
+                    (
+                        bob_address.to_string(),
+                        json!({
+                        "coin_balance": amount_to_json((bob_balance + bob_locked_balance).unwrap(), chain_config.coin_decimals()),
+                        "locked_coin_balance": amount_to_json(Amount::ZERO, chain_config.coin_decimals()),
+                        "transaction_history": bob_transaction_history,
+                                }),
+                    ),
+                ]);
+
+                chainstate_block_ids
+                    .iter()
+                    .map(|id| tf.block(tf.to_chain_block_id(id.into())))
+                    .collect::<Vec<_>>()
+            };
+
+            let storage = {
+                let mut storage = TransactionalApiServerInMemoryStorage::new(&chain_config);
+
+                let mut db_tx = storage.transaction_rw().await.unwrap();
+                db_tx.initialize_storage(&chain_config).await.unwrap();
+                db_tx.commit().await.unwrap();
+
+                storage
+            };
+
+            let chain_config = Arc::new(chain_config);
+
+            let mut local_node = BlockchainState::new(Arc::clone(&chain_config), storage);
+            local_node.scan_genesis(chain_config.genesis_block()).await.unwrap();
+            local_node.scan_blocks(BlockHeight::new(0), chainstate_blocks).await.unwrap();
+
+            ApiServerWebServerState {
+                db: Arc::new(local_node.storage().clone_storage().await),
+                chain_config: Arc::clone(&chain_config),
+                rpc: Arc::new(DummyRPC {}),
+                cached_values: Arc::new(CachedValues {
+                    feerate_points: RwLock::new((get_time(), vec![])),
+                }),
+                time_getter: Default::default(),
+            }
+        };
+
+        web_server(listener, web_server_state, true).await
+    });
+
+    for (address, expected_balance) in rx.await.unwrap() {
+        let url = format!("/api/v2/address/{address}");
+
+        // Given that the listener port is open, this will block until a
+        // response is made (by the web server, which takes the listener
+        // over)
+        let response = reqwest::get(format!("http://{}:{}{url}", addr.ip(), addr.port()))
+            .await
+            .unwrap();
+
+        assert_eq!(
+            response.status(),
+            200,
+            "Failed getting address balance for {address}"
+        );
+
+        let body = response.text().await.unwrap();
+        let body: serde_json::Value = serde_json::from_str(&body).unwrap();
+
+        assert_eq!(body, expected_balance);
+    }
+
+    task.abort();
+}
+
+#[rstest]
+#[trace]
+#[case(Seed::from_entropy())]
+#[tokio::test]
+async fn ok(#[case] seed: Seed) {
+    let listener = TcpListener::bind("127.0.0.1:0").unwrap();
+    let addr = listener.local_addr().unwrap();
+
+    let (tx, rx) = tokio::sync::oneshot::channel();
+
+    let task = tokio::spawn(async move {
+        let web_server_state = {
+            let mut rng = make_seedable_rng(seed);
+            let chain_config = create_unit_test_config();
+
+            let chainstate_blocks = {
+                let mut tf = TestFramework::builder(&mut rng)
+                    .with_chain_config(chain_config.clone())
+                    .build();
+
+                // generate addresses
+
+                let (alice_sk, alice_pk) =
+                    PrivateKey::new_from_rng(&mut rng, KeyKind::Secp256k1Schnorr);
+
+                let alice_destination = Destination::PublicKeyHash(PublicKeyHash::from(&alice_pk));
+                let alice_address =
+                    Address::<Destination>::new(&chain_config, &alice_destination).unwrap();
+                let mut alice_balance = Amount::from_atoms(1_000_000);
+                let mut alice_transaction_history: Vec<Id<Transaction>> = vec![];
+
+                let (_bob_sk, bob_pk) =
+                    PrivateKey::new_from_rng(&mut rng, KeyKind::Secp256k1Schnorr);
+
+                let bob_destination = Destination::PublicKeyHash(PublicKeyHash::from(&bob_pk));
+                let bob_address =
+                    Address::<Destination>::new(&chain_config, &bob_destination).unwrap();
+                let mut bob_balance = Amount::ZERO;
+                let mut bob_transaction_history: Vec<Id<Transaction>> = vec![];
+
+                // setup initial transaction
+
+                let mut previous_tx_out =
+                    TxOutput::Transfer(OutputValue::Coin(alice_balance), alice_destination.clone());
+
+                let transaction = TransactionBuilder::new()
+                    .add_input(
+                        TxInput::from_utxo(
+                            OutPointSourceId::BlockReward(tf.genesis().get_id().into()),
+                            0,
+                        ),
+                        InputWitness::NoSignature(None),
+                    )
+                    .add_output(previous_tx_out.clone())
+                    .build();
+
+                alice_transaction_history.push(transaction.transaction().get_id());
+                let mut previous_transaction_id = transaction.transaction().get_id();
+
+                let mut previous_witness = InputWitness::Standard(
+                    StandardInputSignature::produce_uniparty_signature_for_input(
+                        &alice_sk,
+                        SigHashType::try_from(SigHashType::ALL).unwrap(),
+                        alice_destination.clone(),
+                        &transaction,
+                        &[Some(&previous_tx_out)],
+                        0,
+                    )
+                    .unwrap(),
+                );
+
+                let mut chainstate_block_ids = vec![*tf
+                    .make_block_builder()
+                    .add_transaction(transaction.clone())
+                    .build_and_process()
+                    .unwrap()
+                    .unwrap()
+                    .block_id()];
+
+                for _ in 0..rng.gen_range(1..100) {
                     let random_coin_amount = rng.gen_range(1..10);
 
                     alice_balance =
-                        (alice_balance - Amount::from_atoms(random_coin_amount * 2)).unwrap();
+                        (alice_balance - Amount::from_atoms(random_coin_amount)).unwrap();
+
+                    bob_balance = (bob_balance + Amount::from_atoms(random_coin_amount)).unwrap();
 
                     let alice_tx_out = TxOutput::Transfer(
                         OutputValue::Coin(alice_balance),
@@ -411,11 +636,6 @@ async fn ok(#[case] seed: Seed) {
                     let bob_tx_out = TxOutput::Transfer(
                         OutputValue::Coin(Amount::from_atoms(random_coin_amount)),
                         bob_destination.clone(),
-                    );
-                    let bob_tx_out2 = TxOutput::LockThenTransfer(
-                        OutputValue::Coin(Amount::from_atoms(random_coin_amount)),
-                        bob_destination.clone(),
-                        OutputTimeLock::ForBlockCount(rng.gen_range(1..100)),
                     );
 
                     let transaction = TransactionBuilder::new()
@@ -428,30 +648,10 @@ async fn ok(#[case] seed: Seed) {
                         )
                         .add_output(alice_tx_out.clone())
                         .add_output(bob_tx_out.clone())
-                        .add_output(bob_tx_out2.clone())
                         .build();
 
-                    alice_utxos.insert(
-                        UtxoOutPoint::new(
-                            OutPointSourceId::Transaction(transaction.transaction().get_id()),
-                            0,
-                        ),
-                        alice_tx_out.clone(),
-                    );
-                    bob_utxos.insert(
-                        UtxoOutPoint::new(
-                            OutPointSourceId::Transaction(transaction.transaction().get_id()),
-                            1,
-                        ),
-                        bob_tx_out,
-                    );
-                    bob_utxos.insert(
-                        UtxoOutPoint::new(
-                            OutPointSourceId::Transaction(transaction.transaction().get_id()),
-                            2,
-                        ),
-                        bob_tx_out2,
-                    );
+                    alice_transaction_history.push(transaction.transaction().get_id());
+                    bob_transaction_history.push(transaction.transaction().get_id());
                     previous_transaction_id = transaction.transaction().get_id();
 
                     previous_witness = InputWitness::Standard(
@@ -487,27 +687,19 @@ async fn ok(#[case] seed: Seed) {
                 _ = tx.send([
                     (
                         alice_address.get().to_string(),
-                        alice_utxos
-                            .into_iter()
-                            .map(|utxo| {
-                                json!({
-                                "outpoint": utxo.0,
-                                "utxo": txoutput_to_json(&utxo.1, &chain_config, &TokenDecimals::Single(None))})
-                            })
-                            .collect::<Vec<_>>()
-                            .into(),
+                        json!({
+                        "coin_balance": amount_to_json(alice_balance, chain_config.coin_decimals()),
+                        "locked_coin_balance": amount_to_json(Amount::ZERO, chain_config.coin_decimals()),
+                        "transaction_history": alice_transaction_history,
+                                }),
                     ),
                     (
                         bob_address.to_string(),
-                        bob_utxos
-                            .into_iter()
-                            .map(|utxo| {
-                                json!({
-                                "outpoint": utxo.0,
-                                "utxo": txoutput_to_json(&utxo.1, &chain_config, &TokenDecimals::Single(None))})
-                            })
-                            .collect::<Vec<_>>()
-                            .into(),
+                        json!({
+                        "coin_balance": amount_to_json(bob_balance, chain_config.coin_decimals()),
+                        "locked_coin_balance": amount_to_json(Amount::ZERO, chain_config.coin_decimals()),
+                        "transaction_history": bob_transaction_history,
+                                }),
                     ),
                 ]);
 
@@ -548,7 +740,7 @@ async fn ok(#[case] seed: Seed) {
     });
 
     for (address, expected_values) in rx.await.unwrap() {
-        let url = format!("/api/v1/address/{address}/all-utxos");
+        let url = format!("/api/v2/address/{address}");
 
         // Given that the listener port is open, this will block until a
         // response is made (by the web server, which takes the listener
