+++ api-server/stack-test-suite/tests/v2/chain_at_height.rs
@@ -23,58 +23,49 @@ use crate::DummyRPC;
 use super::*;
 
 #[tokio::test]
-async fn at_genesis() {
-    let url = "/api/v1/chain/tip";
+async fn invalid_height() {
+    let (task, response) = spawn_webserver("/api/v2/chain/invalid-height").await;
 
-    let listener = TcpListener::bind("127.0.0.1:0").unwrap();
-    let addr = listener.local_addr().unwrap();
-
-    let (tx, rx) = tokio::sync::oneshot::channel();
-
-    let task = tokio::spawn({
-        async move {
-            let web_server_state = {
-                let chain_config = Arc::new(create_unit_test_config());
-                let binding = Arc::clone(&chain_config);
-                let expected_genesis_id = binding.genesis_block().get_id();
-
-                _ = tx.send(json!({
-                    "block_height": 0,
-                    "block_id": expected_genesis_id.to_hash().encode_hex::<String>(),
-                }));
-
-                let storage = TransactionalApiServerInMemoryStorage::new(&chain_config);
-
-                ApiServerWebServerState {
-                    db: Arc::new(storage),
-                    chain_config: chain_config.clone(),
-                    rpc: Arc::new(DummyRPC {}),
-                    cached_values: Arc::new(CachedValues {
-                        feerate_points: RwLock::new((get_time(), vec![])),
-                    }),
-                    time_getter: Default::default(),
-                }
-            };
-
-            web_server(listener, web_server_state, true).await
-        }
-    });
-
-    // Given that the listener port is open, this will block until a
-    // response is made (by the web server, which takes the listener
-    // over)
-    let response = reqwest::get(format!("http://{}:{}{url}", addr.ip(), addr.port()))
-        .await
-        .unwrap();
-
-    assert_eq!(response.status(), 200);
+    assert_eq!(response.status(), 400);
 
     let body = response.text().await.unwrap();
     let body: serde_json::Value = serde_json::from_str(&body).unwrap();
 
-    let expected_tip = rx.await.unwrap();
+    assert_eq!(body["error"].as_str().unwrap(), "Invalid block height");
 
-    assert_eq!(body, expected_tip);
+    task.abort();
+}
+
+#[tokio::test]
+async fn height_zero() {
+    let (task, response) = spawn_webserver("/api/v2/chain/0").await;
+
+    assert_eq!(response.status(), 404);
+
+    let body = response.text().await.unwrap();
+    let body: serde_json::Value = serde_json::from_str(&body).unwrap();
+
+    assert_eq!(
+        body["error"].as_str().unwrap(),
+        "No block found at supplied height"
+    );
+
+    task.abort();
+}
+
+#[tokio::test]
+async fn height_past_tip() {
+    let (task, response) = spawn_webserver("/api/v2/chain/1337").await;
+
+    assert_eq!(response.status(), 404);
+
+    let body = response.text().await.unwrap();
+    let body: serde_json::Value = serde_json::from_str(&body).unwrap();
+
+    assert_eq!(
+        body["error"].as_str().unwrap(),
+        "No block found at supplied height"
+    );
 
     task.abort();
 }
@@ -84,7 +75,10 @@ async fn at_genesis() {
 #[case(Seed::from_entropy())]
 #[tokio::test]
 async fn height_n(#[case] seed: Seed) {
-    let url = "/api/v1/chain/tip";
+    let mut rng = make_seedable_rng(seed);
+    let block_height = rng.gen_range(1..50);
+    let n_blocks = rng.gen_range(block_height..100);
+    let url = format!("/api/v2/chain/{block_height}");
 
     let listener = TcpListener::bind("127.0.0.1:0").unwrap();
     let addr = listener.local_addr().unwrap();
@@ -93,9 +87,6 @@ async fn height_n(#[case] seed: Seed) {
 
     let task = tokio::spawn({
         async move {
-            let mut rng = make_seedable_rng(seed);
-            let n_blocks = rng.gen_range(1..100);
-
             let web_server_state = {
                 let chain_config = create_unit_test_config();
 
@@ -109,12 +100,9 @@ async fn height_n(#[case] seed: Seed) {
                         .unwrap();
 
                     // Need the "- 1" to account for the genesis block not in the vec
-                    let expected_block_id = chainstate_block_ids[n_blocks - 1];
+                    let expected_block_id = chainstate_block_ids[block_height - 1];
 
-                    _ = tx.send(json!({
-                    "block_height": n_blocks,
-                    "block_id": expected_block_id,
-                    }));
+                    _ = tx.send(expected_block_id);
 
                     chainstate_block_ids
                         .iter()
@@ -164,9 +152,12 @@ async fn height_n(#[case] seed: Seed) {
     let body = response.text().await.unwrap();
     let body: serde_json::Value = serde_json::from_str(&body).unwrap();
 
-    let expected_tip = rx.await.unwrap();
+    let expected_block_id = rx.await.unwrap();
 
-    assert_eq!(body, expected_tip);
+    assert_eq!(
+        body.as_str().unwrap(),
+        expected_block_id.to_hash().encode_hex::<String>()
+    );
 
     task.abort();
 }
