+++ api-server/stack-test-suite/tests/v2/nft.rs
@@ -25,7 +25,7 @@ use super::*;
 
 #[tokio::test]
 async fn invalid_nft_id() {
-    let (task, response) = spawn_webserver("/api/v1/nft/invalid-token-id").await;
+    let (task, response) = spawn_webserver("/api/v2/nft/invalid-token-id").await;
 
     assert_eq!(response.status(), 400);
 
@@ -48,7 +48,7 @@ async fn nft_not_found(#[case] seed: Seed) {
     let token_id = TokenId::new(H256::random_using(&mut rng));
     let token_id = Address::<TokenId>::new(&chain_config, &token_id).unwrap();
 
-    let (task, response) = spawn_webserver(&format!("/api/v1/nft/{}", token_id.get())).await;
+    let (task, response) = spawn_webserver(&format!("/api/v2/nft/{}", token_id.get())).await;
 
     assert_eq!(response.status(), 404);
 
@@ -180,7 +180,7 @@ async fn ok(#[case] seed: Seed) {
     let chain_config = create_unit_test_config();
     for (token_id, expected_values) in rx.await.unwrap() {
         let token_id = Address::new(&chain_config, &token_id).unwrap();
-        let url = format!("/api/v1/nft/{token_id}");
+        let url = format!("/api/v2/nft/{token_id}");
 
         // Given that the listener port is open, this will block until a
         // response is made (by the web server, which takes the listener
