+++ api-server/stack-test-suite/tests/v2/pool_block_stats.rs
@@ -24,7 +24,7 @@ use super::*;
 #[tokio::test]
 async fn invalid_pool_id() {
     let (task, response) =
-        spawn_webserver("/api/v1/pool/invalid-transaction-id/block-stats?from=0&to=0").await;
+        spawn_webserver("/api/v2/pool/invalid-transaction-id/block-stats?from=0&to=0").await;
 
     assert_eq!(response.status(), 400);
 
@@ -41,7 +41,7 @@ async fn from_to_not_specified() {
     let pool_id = PoolId::new(H256::zero());
     let chain_config = create_unit_test_config();
     let pool_id = Address::new(&chain_config, &pool_id).unwrap();
-    let (task, response) = spawn_webserver(&format! {"/api/v1/pool/{pool_id}/block-stats"}).await;
+    let (task, response) = spawn_webserver(&format! {"/api/v2/pool/{pool_id}/block-stats"}).await;
 
     assert_eq!(response.status(), 400);
 
@@ -60,7 +60,7 @@ async fn pool_id_not_fund() {
     let chain_config = create_unit_test_config();
     let pool_id = Address::new(&chain_config, &pool_id).unwrap();
     let (task, response) =
-        spawn_webserver(&format! {"/api/v1/pool/{pool_id}/block-stats?from=0&to=0"}).await;
+        spawn_webserver(&format! {"/api/v2/pool/{pool_id}/block-stats?from=0&to=0"}).await;
 
     assert_eq!(response.status(), 404);
 
@@ -176,7 +176,7 @@ async fn ok(#[case] seed: Seed) {
     let chain_config = create_regtest();
     let (pool_id, num_blocks, (from, to)) = rx.await.unwrap();
     let pool_id = Address::new(&chain_config, &pool_id).unwrap();
-    let url = format!("/api/v1/pool/{pool_id}/block-stats?from={from}&to={to}");
+    let url = format!("/api/v2/pool/{pool_id}/block-stats?from={from}&to={to}");
 
     // Given that the listener port is open, this will block until a
     // response is made (by the web server, which takes the listener
