+++ api-server/stack-test-suite/tests/v2/transaction_merkle_path.rs
@@ -27,7 +27,7 @@ use std::str::FromStr;
 
 #[tokio::test]
 async fn get_transaction_failed() {
-    let (task, response) = spawn_webserver("/api/v1/transaction/invalid-txid/merkle-path").await;
+    let (task, response) = spawn_webserver("/api/v2/transaction/invalid-txid/merkle-path").await;
 
     assert_eq!(response.status(), 400);
 
@@ -155,7 +155,7 @@ async fn get_block_failed(#[case] seed: Seed) {
     });
 
     let transaction_id = rx.await.unwrap();
-    let url = format!("/api/v1/transaction/{transaction_id}/merkle-path");
+    let url = format!("/api/v2/transaction/{transaction_id}/merkle-path");
 
     // Given that the listener port is open, this will block until a
     // response is made (by the web server, which takes the listener
@@ -270,7 +270,7 @@ async fn transaction_not_part_of_block(#[case] seed: Seed) {
     });
 
     let transaction_id = rx.await.unwrap();
-    let url = format!("/api/v1/transaction/{transaction_id}/merkle-path");
+    let url = format!("/api/v2/transaction/{transaction_id}/merkle-path");
 
     // Given that the listener port is open, this will block until a
     // response is made (by the web server, which takes the listener
@@ -400,7 +400,7 @@ async fn cannot_find_transaction_in_block(#[case] seed: Seed) {
     });
 
     let transaction_id = rx.await.unwrap();
-    let url = format!("/api/v1/transaction/{transaction_id}/merkle-path");
+    let url = format!("/api/v2/transaction/{transaction_id}/merkle-path");
 
     // Given that the listener port is open, this will block until a
     // response is made (by the web server, which takes the listener
@@ -514,7 +514,7 @@ async fn ok(#[case] seed: Seed) {
     });
 
     let (transaction_id, expected_path) = rx.await.unwrap();
-    let url = format!("/api/v1/transaction/{transaction_id}/merkle-path");
+    let url = format!("/api/v2/transaction/{transaction_id}/merkle-path");
 
     // Given that the listener port is open, this will block until a
     // response is made (by the web server, which takes the listener
